TLS Fixes with self-signed certificate
If dirsrv_selfsigned_cert is set, we shouldn't disable TLS, and should
allow TLS to be enforced by the server.

The consistency check in main.yml will already catch the case where
no TLS is available and can thus not be enforced. So here we can
limit ourselves to check dirsrv_tls_enforced.