main.yml: use LDAPI to connect if possible
LDAPI connections don't use TLS, even if TLS is required on regular sockets.
(At least not if localssf is >= minssf). Therefore if LDAPI is enabled,
it makes sense to use it to connect instead of using TLS over a regular
socket, which is more error-prone (it fails if the certificate has expired, for
example).

Also, rather than running grep over dse.ldif, just read it in and examine it
using native ansible code, which is considerably faster.